% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{AsymDiv}
\alias{AsymDiv}
\title{Asymptotic Diversity main function}
\usage{
AsymDiv(
  x,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 30,
  conf = 0.95,
  method = "Both"
)
}
\arguments{
\item{x}{a matrix/data.frame/list/vector of abundances-based/incidences-based species data.\cr}

\item{q}{a integer vector for the order of Hill number\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).\cr}

\item{nboot}{an integer specifying the number of bootstrap replications, default is 30.\cr}

\item{conf}{positive number < 1 specifying the level of confidence interval, default is 0.95.\cr\cr}

\item{method}{a binary calculation method with 'Estimated' or 'Empirical'\cr}
}
\value{
a matrix of estimated Asymptotic Diversity with order q: \cr\cr
}
\description{
\code{AsymDiv} Estimation of Asymptotic Diversity with order q
}
\examples{
\dontrun{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- AsymDiv(x = Spider, datatype = "abundance")
out1

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_incid)
out2 <- AsymDiv(x = woody_incid[,c(1,4)], datatype = "incidence_freq")
out2
}
}
\references{
Chao,A. and Jost,L.(2015).Estimating diversity and entropy profiles via discovery rates of new species.
}
