\name{Brazil_abun_data}
\alias{Brazil_abun_data}
\docType{data}
\title{
Abundance data (datatype = "abundance")
}
\description{
  This dataset includes tree species abundances in two assemblages/habitats (Edge and Interior) collected from rainforests in Brazil by Magnago et al. (2014, 2015, 2017). \cr
}
\usage{data(Brazil_abun_data)}
\format{
  \code{Brazil_abun_data} is a species-by-assemblages data.frame (in addition to species names and assemblage names) with 425 tree species and two assemblages (Edge and Interior). \cr
  $ Edge: int 11 110 36 12 2 7 11 1 1 0 ... \cr
  $ Interior: int 21 11 7 1 0 1 11 13 1 2 ... \cr
 }
\examples{
data(Brazil_abun_data)
}
\references{
Magnago, L. F. S., Edwards, D. P., Edwards, F. A., Magrach, A., Martins, S. V., & Laurance, W. F. (2014). Functional attributes change but functional richness is unchanged after fragmentation of Brazilian Atlantic forests. Journal of Ecology, 102, 475-485. \cr\cr
Magnago, L. F. S., Magrach, A., Barlow, J., Schaefer, C. E. G. R., Laurance, W. F., Martins, S. V., & Edwards, D. P. (2017). Do fragment size and edge effects predict carbon stocks in trees and lianas in tropical forests? Functional Ecology, 31, 542-552. \cr\cr
Magnago, L. F. S., Rocha, M. F., Meyer, L., Martins, S. V., & Meira-Neto, J. A. A. (2015). Microclimatic conditions at forest edges have significant impacts on vegetation structure in large Atlantic forest fragments. Biodiversity and Conservation, 24, 2305-2318. \cr\cr
}

\keyword{datasets}
