\name{Brazil_phylo_tree}
\alias{Brazil_phylo_tree}
\docType{data}
\title{
Phylogenetic tree for \code{Brazil_rainforest_abun_data}
}
\description{
  This dataset includes a phylogenetic tree spanned by 425 species listed in the dataset \code{Brazil_rainforest_abun_data}. \cr
}
\usage{data(Brazil_phylo_tree)}
\format{
  \code{Brazil_phylo_tree} is a list with the following phylogenetic information: \cr
  A list of 5: \cr
  $ edge: int [1:629, 1:2] 426 427 428 429 430 431 427 428 429 430 ... \cr
  $ edge.length: num [1:629] 75 146 8 6 4 14 10 10 127 6 ... \cr
  $ Nnode: int 205 \cr
  $ node.label: chr [1:205] "magnoliales_to_asterales" "poales_to_asterales" ... \cr
  $ tip.label: chr [1:425] "Carpotroche_brasiliensis" "Casearia_ulmifolia" "Casearia_sp4" ... \cr
 }
\examples{
data(Brazil_phylo_tree)
}
\keyword{datasets}
