% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{Completeness}
\alias{Completeness}
\title{Sample Completeness main function}
\usage{
Completeness(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  conf = 0.95,
  nT = NULL
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_freq"}, data can be input as a vector of incidence frequencies (for a single assemblage), matrix/data.frame (species by assemblages), or a list of incidence frequencies; the first entry in all types of input must be the number of sampling units in each assemblage. \cr
(c) For \code{datatype = "incidence_raw"}, data can be input as a list of matrix/data.frame (species by sampling units); data can also be input as a matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (\code{nT}, see below) must be input.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{(0, 0.2, 0.4,...,2)}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}), or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being \code{0} (non-detection) or \code{1} (detection).}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter \code{0} to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < \code{1} specifying the level of confidence interval. Default is \code{0.95}.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data is matrix/data.frame) a vector of nonnegative integers specifying the number of sampling units in each assemblage. If assemblage names are not specified, then assemblages are automatically named as "assemblage1", "assemblage2",..., etc.}
}
\value{
a matrix of estimated sample completeness with order q: 
        \item{Order.q}{the diversity order of q.}
        \item{Estimate.SC}{the estimated (or observed) sample completeness of order q.}
        \item{s.e.}{standard error of sample completeness.}
        \item{SC.LCL, SC.UCL}{the bootstrap lower and upper confidence limits for the sample completeness of order q at the specified level (with a default value of \code{0.95}).}
        \item{Assemblage}{the assemblage name.}
}
\description{
\code{Completeness} Estimation of Sample Completeness with order q
}
\examples{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
output1 <- Completeness(data = Spider, datatype = "abundance")
output1

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_plants)
output2 <- Completeness(data = woody_plants[,c(1,4)], datatype = "incidence_freq")
output2

}
\references{
Chao, A., Y. Kubota, D. Zelen??, C.-H. Chiu, C.-F. Li, B. Kusumoto, M. Yasuhara, S. Thorn, C.-L. Wei, M. J. Costello, and R. K. Colwell (2020). Quantifying sample completeness and comparing diversities among assemblages. Ecological Research, 35, 292-314.
}
