% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{Diversity_emp}
\alias{Diversity_emp}
\title{Empirical Diversity}
\usage{
Diversity_emp(x, q, datatype)
}
\arguments{
\item{x}{a vector of abundances-based/incidences-based species data.\cr}

\item{q}{a integer vector for the order of Hill number\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).\cr}
}
\value{
a vector of empirical Diversity with order q: \cr\cr
}
\description{
\code{Diversity_emp} Empirical Diversity with order q
}
