% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{ggAsymDiv}
\alias{ggAsymDiv}
\title{ggplot for Asymptotic diversity}
\usage{
ggAsymDiv(output)
}
\arguments{
\item{output}{a table generated from AsymDiv function\cr}
}
\value{
a figure of estimated sample completeness with order q\cr\cr
}
\description{
\code{ggAsymDiv} The figure for estimation of Asymptotic diversity with order q\cr
}
\examples{
\dontrun{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- AsymDiv(x = Spider, datatype = "abundance")
ggAsymDiv(out1)
## Type (2) example for incidence based data (list of data.frame)

## Ex.2
data(woody_incid)
out2 <- AsymDiv(x = woody_incid[,c(1,4)], datatype = "incidence_freq")
ggAsymDiv(out2)
}
}
\references{
Chao,A. and Jost,L.(2015).Estimating diversity and entropy profiles via discovery rates of new species.
}
