% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggCompleteness}
\alias{ggCompleteness}
\title{ggplot for Sample Completeness}
\usage{
ggCompleteness(output)
}
\arguments{
\item{output}{the output of the function \code{Completeness}.}
}
\value{
a figure of estimated sample completeness with order q.
}
\description{
\code{ggCompleteness} is a \code{ggplot2} extension for \code{Completeness} Object to plot sample completeness with order q.\cr\cr
}
\examples{
## Type (1) example for abundance-based data
## Example 1
\donttest{
data(Data_spider)
output1 <- Completeness(data = Data_spider, datatype = "abundance")
ggCompleteness(output1)
}

## Type (2) example for incidence-based data
## Example 2
data(Data_woody_plant)
output2 <- Completeness(data = Data_woody_plant, datatype = "incidence_raw")
ggCompleteness(output2)


}
