% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggEvenness}
\alias{ggEvenness}
\title{ggplot for Evenness
\code{ggEvenness} is a \code{ggplot2} extension for \code{Evenness} Object to plot evenness with order q.\cr\cr}
\usage{
ggEvenness(output)
}
\arguments{
\item{output}{the output of the function \code{Evenness}.\cr}
}
\value{
a figure of estimated (or observed) evenness with order q.\cr
}
\description{
ggplot for Evenness
\code{ggEvenness} is a \code{ggplot2} extension for \code{Evenness} Object to plot evenness with order q.\cr\cr
}
\examples{
## Type (1) example for abundance based data
# Example 1. observed evenness for abundance data
data(Spider)
output1 <- Evenness(data = Spider, datatype = "abundance", 
                    method = "Observed", E.class = 1:5)
ggEvenness(output1)

\donttest{
# Example 2. estimated evenness for abundance data with default SC = Cmax
data(Spider)
output2 <- Evenness(data = Spider, datatype = "abundance", 
                    method = "Estimated", SC = NULL, E.class = 1:5)
ggEvenness(output2)
}

## Type (2) example for incidence based data
# Example 3. observed evenness for incidence data
data(Woody_plants)
output3 <- Evenness(data = Woody_plants[,c(1,4)], datatype = "incidence_freq", 
                    method = "Observed", E.class = 1:5)
ggEvenness(output3)

\donttest{
# Example 4. estimated evenness for incidence data with default SC = Cmax
data(Woody_plants)
output4 <- Evenness(data = Woody_plants[,c(1,4)], datatype = "incidence_freq", 
                    method = "Estimated", SC = NULL, E.class = 1:5)
ggEvenness(output4)
}

}
