% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{ggRE}
\alias{ggRE}
\title{ggplot2 extension for an iNEXT object of Phylogenetic and Functional diversity}
\usage{
ggRE(output)
}
\arguments{
\item{output}{a table generated from iNEXTFD function\cr}
}
\value{
a figure of estimated sample completeness with order q\cr
}
\description{
\code{ggRE} The figure for estimation of rarefaction and extrapolation with order q\cr
}
