% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{ggSC}
\alias{ggSC}
\title{ggplot for Sample Completeness}
\usage{
ggSC(output)
}
\arguments{
\item{output}{a table generated from SC function}
}
\value{
a figure of estimated sample completeness with order q
}
\description{
\code{ggSC} the \code{\link[ggplot2]{ggplot}} extension for \code{\link{SC}} Object to plot sample completeness with order q
}
\examples{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- SC(data = Spider, datatype = "abundance")
ggSC(out1)

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_incid)
out2 <- SC(data = woody_incid[,c(1,4)], datatype = "incidence_freq")
ggSC(out2)

}
\references{
Chao,A.,Y.Kubota,D.Zelený,C.-H.Chiu.
Quantifying sample completeness and comparing diversities among assemblages.
}
