% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{ggtqplotD}
\alias{ggtqplotD}
\title{ggplot for Asymptotic diversity}
\usage{
ggtqplotD(output)
}
\arguments{
\item{output}{a table generated from AsymDiv function\cr}
}
\value{
a figure of estimated sample completeness with order q\cr\cr
}
\description{
\code{ggtqplotD} The figure for estimation of Asymptotic diversity with order q\cr
}
\examples{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- TdAsy(x = Spider, datatype = "abundance")
ggtqplotD(out1)
## Type (2) example for incidence based data (list of data.frame)

## Ex.2
data(woody_incid)
out2 <- TdObs(x = woody_incid[,c(1,4)], datatype = "incidence_freq")
ggtqplotD(out2)

}
\references{
Chao,A. and Jost,L.(2015).Estimating diversity and entropy profiles via discovery rates of new species.
}
