% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{iNEXT.4steps}
\alias{iNEXT.4steps}
\title{iNEXT 4 steps}
\usage{
iNEXT.4steps(
  data,
  datatype = "abundance",
  size = NULL,
  endpoint = NULL,
  knots = 30,
  se = TRUE,
  conf = 0.95,
  nboot = 30,
  details = FALSE
)
}
\arguments{
\item{data}{a matrix/data.frame of species abundances/incidences data.\cr
Type (1) abundance data:\cr
When there are N assemblages, the
observed species abundances should be arranged as a species (in rows) by assemblage (in columns) matrix. The first row
(including N entries) lists the assemblage labels or site names for the N assemblages.\cr
Type (2) incidence data:\cr
The data input format for incidence data must be raw detection/non-detection data. That is, data for each community/assemblage
consist of a species-by-sampling-unit matrix. Users must first merge multiple-community data by species identity to obtain a pooled
list of species; then the rows of the input data refer to this pooled list. \cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).\cr}

\item{size}{a vector of nonnegative integers specifying the sample sizes for which diversity estimates will be calculated. If \code{NULL}, the diversity estimates will
be calculated for those sample sizes determined by the specified/default \code{endpoint} and \code{knot}. \cr
(setting only for \code{step2}).\cr}

\item{endpoint}{an interger specifying the endpoint for rarefaction and extrapolation range. If \code{NULL}, \code{endpoint} = double of the maximum
reference sample size. It will be ignored if \code{size} is given. \cr
(setting only for \code{step2}).\cr}

\item{knots}{an integer specifying the number of knot between 1 and the \code{endpoint}, default is 40.\cr
(setting only for \code{step2}).\cr}

\item{se}{a logical variable to calculate the bootstrap standard error and confidence interval of a level specified by conf, default is \code{TRUE}.\cr
(setting only for \code{step2}).\cr}

\item{conf}{a positive number < 1 specifying the level of confidence interval, default is 0.95.\cr
(setting only for \code{step2}).\cr}

\item{nboot}{an integer specifying the number of bootstrap replications, default is 30.\cr
(setting only for \code{step2}).\cr}

\item{details}{a logical variable to determine whether do you want to print out the detailed value of 4 plots, default is \code{FALSE}.\cr}
}
\value{
a list of two of objects: \cr\cr
\code{$summary} individual summary of 4 steps of data. \cr\cr
\code{$figure} 5 figures of analysis process. \cr\cr
\code{$details} the information for generating \code{figure}. \cr
if you nees it, you should key in \code{details = TRUE}. \cr\cr
}
\description{
\code{iNEXT.4steps}:\cr
A complete (random sampling) biological analysis combined with four parts:\cr
Step1: Sample Completeness.\cr
Step2: Interpolation and Extrapolation.\cr
Step3: Asymptotic diversity.\cr
Step4: Evenness.\cr
}
\examples{
\dontrun{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(spider)
out1 <- iNEXT.4steps(data = spider, datatype = "abundance")
out1
## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_incid)
out2 <- iNEXT.4steps(data = woody_incid[,c(1,4)], datatype = "incidence_freq")
out2
}
}
\references{
Chao,A., Y.Kubota, D.Zeleny, C.-H.Chiu, C.-F.Li, B.Kusumoto, M.Yasuhara, S.Thorn, C.-L.Wei, M.J.Costello, and R.K.olwell(2020).
Quantifying sample completeness and comparing diversities among assemblages. Ecological Research.
}
