% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{iNEXT4steps}
\alias{iNEXT4steps}
\title{iNEXT 4 steps}
\usage{
iNEXT4steps(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 30,
  conf = 0.95,
  nT = NULL,
  details = FALSE
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_raw"}, data can be input as a list of matrix/data.frame (species by sampling units); data can also be input as a matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (nT, see below) must be input.}

\item{q}{a numerical vector specifying the diversity orders for q-profile output. Default is \code{seq(0, 2, by = 0.2)}.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 30.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data in a single matrix/data.frame) a vector of positive integers specifying the number of sampling units in each assemblage. If assemblage names are not specified (i.e., \code{names(nT) = NULL}), then assemblages are automatically named as "assemblage1", "assemblage2",..., etc.}

\item{details}{a logical variable to decide whether print out the detailed values for each steps or not. Default is \code{FALSE}.}
}
\value{
a list of three of objects: \cr\cr
\code{$summary} four tables for individual summary of 4 steps. \cr\cr
\code{$figure} six figures (five figures of each analysis process and an overall figure). \cr\cr
\code{$details} (only when \code{details = TRUE}). The numerical output for plotting \code{figure}. \cr\cr
}
\description{
\code{iNEXT4steps}:\cr
A complete (random sampling) biological analysis combined with four parts:\cr
Step 1. Sample completeness profiles\cr
Step 2. Asymptotic analysis\cr
Step 3. Non-asymptotic coverage-based rarefaction and extrapolation\cr
Step 4. Evenness among species abundances\cr
}
\examples{
\donttest{
## Taxonomic diversity for abundance data
data(Data_spider)
output1 <- iNEXT4steps(data = Data_spider, datatype = "abundance")
output1


## Taxonomic diversity for incidence data
data(Data_woody_plant)
output2 <- iNEXT4steps(data = Data_woody_plant[c(1,4)], datatype = "incidence_raw")
output2
}


}
\references{
Chao, A., Y. Kubota, D. Zeleny, C.-H. Chiu, C.-F. Li, B. Kusumoto, M. Yasuhara, S. Thorn, C.-L. Wei, M. J. Costello, and R. K. Colwell (2020). Quantifying sample completeness and comparing diversities among assemblages. Ecological Research, 35, 292-314.
}
