% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{SC}
\alias{SC}
\title{Sample Completeness main function}
\usage{
SC(x, q = seq(0, 2, 0.2), datatype = "abundance", nboot = 30, conf = 0.95)
}
\arguments{
\item{x}{a matrix/data.frame/list/vector of abundances-based/incidences-based species data.\cr}

\item{q}{a integer vector for the order of Hill number\cr}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).\cr}

\item{nboot}{an integer specifying the number of bootstrap replications, default is 30.\cr}

\item{conf}{positive number < 1 specifying the level of confidence interval, default is 0.95.\cr\cr}
}
\value{
a matrix of estimated sample completeness with order q: \cr\cr
}
\description{
\code{SC} Estimation of Sample Completeness with order q
}
\examples{
\dontrun{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- SC(x = Spider, datatype = "abundance")
out1

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_incid)
out2 <- SC(x = woody_incid[,c(1,4)], datatype = "incidence_freq")
out2
}
}
\references{
Chao,A.,Y.Kubota,D.Zelený,C.-H.Chiu.
Quantifying sample completeness and comparing diversities among assemblages.
}
