% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{Evenness}
\alias{Evenness}
\title{Evenness main function}
\usage{
Evenness(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  method = "Estimated",
  nboot = 50,
  conf = 0.95,
  nT = NULL,
  E.class = 1:5,
  C = NULL
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_freq"}, data can be input as a vector of incidence frequencies (for a single assemblage), matrix/data.frame (species by assemblages), or a list of incidence frequencies; the first entry in all types of input must be the number of sampling units in each assemblage. \cr
(c) For \code{datatype = "incidence_raw"}, data can be input as a list of matrix/data.frame (species by sampling units); data can also be input as a matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (nT, see below) must be input.}

\item{q}{a numerical vector specifying the diversity orders. Default is (0, 0.2, 0.4,...,2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}), or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{method}{a binary calculation method with 'Estimated' or 'Observed'.\cr}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data is matrix/data.frame) a vector of nonnegative integers specifying the number of sampling units in each assemblage. If assemblage names are not specified, then assemblages are automatically named as "assemblage1", "assemblage2",..., etc.}

\item{E.class}{an integer vector between 1 to 5}

\item{C}{a standardized coverage for calculating estimated evenness. It is used when \code{method = 'Estimated'}. If \code{NULL}, then this function computes the diversity estimates for the minimum sample coverage among all samples extrapolated to double reference sizes (C = Cmax).}
}
\value{
A list of estimated(observed) evenness with order q.\cr
        Different lists represent different classes of Evenness.\cr
        Each list is combined with order.q and sites.\cr
        If "method" is estimated, then fist list will be named "C" which means the
        maximum standardized coverage among all double reference sample size.\cr\cr
\code{$summary} individual summary of 4 steps of data. \cr\cr
}
\description{
\code{Evenness} Estimation (Observed) of Evenness with order q
}
\details{
R scipts "Evenness" for Chao and Ricotta (2019) Ecology paper.
This R code is for computing Figures 2, 3 and 4 of Chao and Ricotta (2019) paper.
installed and loaded before running the scripts.
}
\examples{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- Evenness(data = Spider, datatype = "abundance")
out1

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_plants)
out2 <- Evenness(data = woody_plants[,c(1,4)], datatype = "incidence_freq")
out2

}
\references{
Chao,A.and Ricotta,C.(2019).Quantifying evenness and linking it to diversity, beta diversity, and similarity.
}
