% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggCompleteness}
\alias{ggCompleteness}
\title{ggplot for Sample Completeness}
\usage{
ggCompleteness(output)
}
\arguments{
\item{output}{the output of the function \code{Completeness}.}
}
\value{
a figure of estimated sample completeness with order q.
}
\description{
\code{ggCompleteness} is a \code{ggplot2} extension for \code{Completeness} Object to plot sample completeness with order q.\cr\cr
}
\examples{
## Type (1) example for abundance-based data
## Example 1
\donttest{
data(Spider)
output1 <- Completeness(data = Spider, datatype = "abundance")
ggCompleteness(output1)
}

## Type (2) example for incidence-based data
## Example 2
data(Woody_plants)
output2 <- Completeness(data = Woody_plants[,c(1,4)], datatype = "incidence_freq")
ggCompleteness(output2)


}
