% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggEvenness}
\alias{ggEvenness}
\title{ggplot for Evenness
\code{ggEvenness} is a \code{ggplot2} extension for \code{Evenness} Object to plot evenness with order q.\cr\cr}
\usage{
ggEvenness(output)
}
\arguments{
\item{output}{the output of the function \code{Evenness}.\cr}
}
\value{
a figure of estimated (or observed) evenness with order q.\cr
}
\description{
ggplot for Evenness
\code{ggEvenness} is a \code{ggplot2} extension for \code{Evenness} Object to plot evenness with order q.\cr\cr
}
\examples{
\donttest{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
output1 <- Evenness(data = Spider, datatype = "abundance", 
                    method = "Estimated", SC = NULL, E.class = 1:5)
ggEvenness(output1)
}

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(Woody_plants)
output2 <- Evenness(data = Woody_plants[,c(1,4)], datatype = "incidence_freq", 
                    method = "Estimated", SC = NULL, E.class = 1:5)
ggEvenness(output2)


}
