% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{ggEvenness}
\alias{ggEvenness}
\title{ggplot for Evenness
\code{ggEvenness} the \code{\link[ggplot2]{ggplot}} extension for \code{\link{Evenness}} Object to plot evenness with order q\cr}
\usage{
ggEvenness(output)
}
\arguments{
\item{output}{a table generated from Evenness function\cr}
}
\value{
a figure of estimated evenness with order q\cr
}
\description{
ggplot for Evenness
\code{ggEvenness} the \code{\link[ggplot2]{ggplot}} extension for \code{\link{Evenness}} Object to plot evenness with order q\cr
}
\examples{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- Evenness(data = Spider, datatype = "abundance")
ggEvenness(out1)

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_plants)
out2 <- Evenness(data = woody_plants[,c(1,4)], datatype = "incidence_freq")
ggEvenness(out2)

}
\references{
Chao,A.and Ricotta,C.(2019).Quantifying evenness and linking it to diversity, beta diversity, and similarity.
}
