% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{iNEXT4steps}
\alias{iNEXT4steps}
\title{iNEXT 4 steps}
\usage{
iNEXT4steps(
  data,
  diversity = c("TD", "PD", "FD"),
  datatype = "abundance",
  nT = NULL,
  PDtree = NULL,
  PDtype = "meanPD",
  FDdistM = NULL,
  FDtype = "AUC",
  nboot = 30,
  p_row = 2,
  p_col = 3,
  details = FALSE
)
}
\arguments{
\item{data}{a matrix/data.frame/list/vector of abundances-based/incidences-based species data.\cr
Type (1) abundance data:\cr
When there are N assemblages, the
observed species abundances should be arranged as a species (in rows) by assemblage (in columns) matrix. The first row
(including N entries) lists the assemblage labels or site names for the N assemblages.\cr
Type (2) incidence data:\cr
The data input format for incidence data must be raw detection/non-detection data. That is, data for each community/assemblage
consist of a species-by-sampling-unit matrix. Users must first merge multiple-community data by species identity to obtain a pooled
list of species; then the rows of the input data refer to this pooled list. \cr}

\item{diversity}{a choice of three-level diversity: 
'TD' = 'Taxonomic', 'PD' = 'Phylogenetic', and 'FD' = 'Functional'.}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}),
sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}) or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}).\cr}

\item{nT}{needed only when \code{datatype = "incidence_raw"}, a sequence of named nonnegative integers specifying the number of sampling units in each assemblage.
If \code{names(nT) = NULL}, then assemblage are automatically named as "assemblage1", "assemblage2",..., etc.
It is necessary when \code{diversity = 'PD'} and \code{datatype = "incidence_raw"}.}

\item{PDtree}{a phylo object describing the phylogenetic tree in Newick format for all observed species in the pooled assemblage. It is necessary when \code{diversity = 'PD'}.}

\item{PDtype}{desired phylogenetic diversity type: PDtype = "PD" for Chao et al. (2010) phylogenetic diversity and PDtype = "meanPD" for mean phylogenetic diversity (phylogenetic Hill number). It will be use when diversity = 'PD'. Default is "meanPD".}

\item{FDdistM}{a pair wise distance matrix for all pairs of observed species in the pooled assemblage. It will be use when \code{diversity = 'FD'}.}

\item{FDtype}{a binary selection for functional type. FDtype = "single" computes diversity under certain threshold. FDtype = "AUC" computes diversity which integrates several threshold between zero and one to get diversity. Default is "AUC".
When you select 'single', then the threshold will be dmean.}

\item{nboot}{an integer specifying the number of bootstrap replications, default is 30.\cr}

\item{p_row}{number of row for 4steps figure, default = 2.}

\item{p_col}{number of column for 4steps figure, default = 3.}

\item{details}{a logical variable to determine whether do you want to print out the detailed value of 4 plots, default is \code{FALSE}.\cr}
}
\value{
a list of two of objects: \cr\cr
\code{$summary} individual summary of 4 steps of data. \cr\cr
\code{$figure} 5 figures of analysis process. \cr\cr
\code{$details} the information for generating \code{figure}. \cr
If you need it, you should key in \code{details = TRUE}. \cr\cr
}
\description{
\code{iNEXT4steps}:\cr
A complete (random sampling) biological analysis combined with four parts:\cr
Step1: Sample Completeness.\cr
Step2: Interpolation and Extrapolation.\cr
Step3: Asymptotic diversity.\cr
Step4: Evenness.\cr
}
\examples{
\dontrun{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- iNEXT4steps(data = Spider, diversity = "TD", datatype = "abundance")
out1

## Ex.2
data(brazil)
data(brazil_tree)
out2 <- iNEXT4steps(data = brazil, diversity = "PD", datatype = "abundance", PDtree = tree, nboot = 0)
out2

## Ex.3
data(brazil)
data(brazil_distM)
out3 <- iNEXT4steps(data = brazil, diversity = "FD", datatype = "abundance", FDdistM = distM, FDtype = 'single', nboot = 0)
out3

## Type (2) example for incidence based data (list of data.frame)
## Ex.1
data(woody_incid)
out <- iNEXT4steps(data = woody_incid[,c(1,4)], diversity = "TD", datatype = "incidence_freq")
out

}

}
\references{
Chao,A., Y.Kubota, D.Zeleny, C.-H.Chiu.
Quantifying sample completeness and comparing diversities among assemblages. Ecological Research.
}
