% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_function.R
\name{iNEXT4steps}
\alias{iNEXT4steps}
\title{iNEXT 4 steps}
\usage{
iNEXT4steps(
  data,
  diversity = "TD",
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 50,
  nT = NULL,
  PDtree = NULL,
  PDreftime = NULL,
  PDtype = "meanPD",
  FDdistM = NULL,
  FDtype = "AUC",
  FDtau = NULL,
  p_row = 2,
  p_col = 3,
  details = FALSE
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_freq"}, data can be input as a vector of incidence frequencies (for a single assemblage), matrix/data.frame (species by assemblages), or a list of incidence frequencies; the first entry in all types of input must be the number of sampling units in each assemblage. \cr
(c) For \code{datatype = "incidence_raw"}, data can be input as a list of matrix/data.frame (species by sampling units); data can also be input as a matrix/data.frame by merging all sampling units across assemblages based on species identity; in this case, the number of sampling units (nT, see below) must be input.}

\item{diversity}{selection of diversity type: \code{'TD'} = Taxonomic diversity, \code{'PD'} = Phylogenetic diversity, and \code{'FD'} = Functional diversity.}

\item{q}{a numerical vector specifying the diversity orders. Default is (0, 0.2, 0.4,...,2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}), or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection).}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data is matrix/data.frame) a vector of nonnegative integers specifying the number of sampling units in each assemblage. If assemblage names are not specified, then assemblages are automatically named as "assemblage1", "assemblage2",..., etc.}

\item{PDtree}{(required only when \code{diversity = "PD"}), a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{PDreftime}{(required only when \code{diversity = "PD"}), a vector of numerical values specifying reference times for PD. Default is \code{NULL} (i.e., the age of the root of PDtree).}

\item{PDtype}{(required only when \code{diversity = "PD"}), select PD type: \code{PDtype = "PD"} (effective total branch length) or \code{PDtype = "meanPD"} (effective number of equally divergent lineages). Default is \code{"meanPD"}, where \code{meanPD = PD/tree depth}.}

\item{FDdistM}{(required only when \code{diversity = "FD"}), a species pairwise distance matrix for all species in the pooled assemblage.}

\item{FDtype}{(required only when \code{diversity = "FD"}), select FD type: \code{FDtype = "tau_values"} for FD under specified threshold values, or \code{FDtype = "AUC"} (area under the curve of tau-profile) for an overall FD which integrates all threshold values between zero and one. Default is \code{"AUC"}.}

\item{FDtau}{(required only when \code{diversity = "FD"} and \code{FDtype = "tau_values"}), a numerical vector between 0 and 1 specifying tau values (threshold levels). If \code{NULL} (default), then threshold is set to be the mean distance between any two individuals randomly selected from the pooled assemblage (i.e., quadratic entropy).}

\item{p_row}{row number for 4 steps figure, default = 2.}

\item{p_col}{column number for 4 steps figure, default = 3.}

\item{details}{a logical variable to decide whether do you want to print out the detailed value for each plots, default is \code{FALSE}.}
}
\value{
a list of three of objects: \cr\cr
\code{$summary} individual summary of 4 steps of data. \cr\cr
\code{$figure} 5 figures of analysis process. \cr\cr
\code{$details} the information for generating \code{figure}. \cr
If you need it, you should key in \code{details = TRUE}. \cr\cr
}
\description{
\code{iNEXT4steps}:\cr
A complete (random sampling) biological analysis combined with four parts:\cr
Step 1. Sample completeness profiles\cr
Step 2. Asymptotic analysis\cr
Step 3. Non-asymptotic coverage-based rarefaction and extrapolation\cr
Step 4. Evenness among species abundances\cr
}
\examples{
\dontrun{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- iNEXT4steps(data = Spider, diversity = "TD", datatype = "abundance")
out1

## Ex.2
data(brazil)
data(brazil_tree)
out2 <- iNEXT4steps(data = brazil, diversity = "PD", datatype = "abundance", 
                    nboot = 0, PDtree = brazil_tree)
out2

## Ex.3
data(brazil)
data(brazil_distM)
out3 <- iNEXT4steps(data = brazil, diversity = "FD", datatype = "abundance", 
                    nboot = 0, FDdistM = brazil_distM, FDtype = 'tau_values')
out3

## Type (2) example for incidence based data (list of data.frame)
## Ex.1
data(woody_incid)
out <- iNEXT4steps(data = woody_incid[,c(1,4)], diversity = "TD", datatype = "incidence_freq")
out

}

}
\references{
Chao,A., Y.Kubota, D.Zeleny, C.-H.Chiu.
Quantifying sample completeness and comparing diversities among assemblages. Ecological Research.
}
