% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_function.R
\name{SC}
\alias{SC}
\title{Sample Completeness main function}
\usage{
SC(
  data,
  q = seq(0, 2, 0.2),
  datatype = "abundance",
  nboot = 30,
  conf = 0.95,
  nT = NULL
)
}
\arguments{
\item{data}{(a) For \code{datatype = "abundance"}, data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors. \cr
(b) For \code{datatype = "incidence_freq"}, data can be input as a vector of incidence frequencies (for a single assemblage), matrix/data.frame (species by assemblages), or a list of incidence frequencies; the first entry in all types of input must be the number of sampling units in each assemblage. \cr
(c) For \code{datatype = "incidence_raw"}, data can be input as a list of matrix/data.frame (species by sampling units); data can also be input as a matrix/data.frame by merging all sampling units across assemblages based on species identity; in this}

\item{q}{a numerical vector specifying the diversity orders. Default is seq(0, 2, by = 0.2).}

\item{datatype}{data type of input data: individual-based abundance data (\code{datatype = "abundance"}), sampling-unit-based incidence frequencies data (\code{datatype = "incidence_freq"}), or species by sampling-units incidence matrix (\code{datatype = "incidence_raw"}) with all entries being 0 (non-detection) or 1 (detection)}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}

\item{nT}{(required only when \code{datatype = "incidence_raw"} and input data is matrix/data.frame) a vector of nonnegative integers specifying the number of sampling units in each assemblage. If assemblage names are not specified, then assemblages are automatically named as "assemblage1", "assemblage2",..., etc.}
}
\value{
a matrix of estimated sample completeness with order q: \cr\cr
}
\description{
\code{SC} Estimation of Sample Completeness with order q
}
\examples{
## Type (1) example for abundance based data (data.frame)
## Ex.1
data(Spider)
out1 <- SC(data = Spider, datatype = "abundance")
out1

## Type (2) example for incidence based data (list of data.frame)
## Ex.2
data(woody_incid)
out2 <- SC(data = woody_incid[,c(1,4)], datatype = "incidence_freq")
out2

}
\references{
Chao,A.,Y.Kubota,D.Zelený,C.-H.Chiu.
Quantifying sample completeness and comparing diversities among assemblages.
}
